wip refactor db
