+++ src/lib/server/dbRoutes/index.ts
@@ -1,0 +1,7 @@
+import createHousehold from './createHousehold';
+import createKid from './createKid';
+
+export {
+    createHousehold,
+    createKid
+}
