+++ tests/calendar.spec.ts
@@ -1,0 +1,13 @@
+/*
+1. mark as busy
+2. mark as unspecified
+3. mark as valid available
+4. mark as invalid available
+5. updating availRange in editor changes displayed editor value
+
+1. correct row colors
+2. render rows properly
+    a. UNSPECIFIED? - Mark Busy btn
+    b. BUSY? - Clear btn
+    c. AVAILABLE? 
+*/
